setup custom colors
kevin provided 4 custom colors (not including white).  I extended the tailwind theme with these 4 colors.  The color names are generic so colors could be changed later and you won't end up with confusing class names.

By extending the provided tailwind colors, I can still use the tailwind colors, which was helpful here because I think Kevin missed a color.  The background of the three column section is not #6F737F....it sure seems lighter to me.

It's obvious some of the text on the dark backgrounds is not pure white, but I don't think it's #6F737F either....that contrast seems off.  I'm going to roll with it for now, and clean it up later if I still don't like it after my styling and layout is done.

tailwind color docs are at https://tailwindcss.com/docs/customizing-colors

specifically for __extending__ the color palette:  https://tailwindcss.com/docs/customizing-colors#adding-additional-colors