setup a custom `max-container`
this allows us to be a little bit DRY.  it also sets a max width for the site w/o setting up a break point.  `container` is a keyword in TW (https://tailwindcss.com/docs/container) and using it would prevent the hero section from being "fluid" across the 700 and 1000px break points we're about to go setup.